Fix OutOfMemoryError when attempting to preserve a large file
Some multi-gigabyte videos and multi-megabyte audio files that were read into memory for creating a preservation file resulted in an OutOfMemoryError. Fixed that by sending a InputStreamProvider that allows opening an InputStream repeatedly, instead of the InputStream itself that was read into the heap for repeated reading resulting in an OOME when the file was very large.