Fix race condition when deleting 2 or more files from an item from which no previous files were deleted
Deleting a file involves creating an archive directory for a record and moving the file from the assetstore folder into the archive folder. If two files are deleted at the same time then the same archive folder is attempted to be created from multiple threads which may lead to a race condition between checking if that directory exists and actually creating it.